SCHED-395: Re-arange components and add abstract layer in-between services
